requirements.txt
streamlit
pydicom
numpy
requests
pillow
matplotlib